Add global namespaced macro for logging.
Created the global namespace macro LOG (dont need to type logging::LogMessage().Get()).
The macro checks if the log level is above the set logging level. If it is, it will not
print the log message. As such, had to change the order of the enums, so debug is the biggest
value.
Created log specific macros (LOG_DEBUG/.../LOG_ERROR) so now one can just write
LOG_ERROR << "My error message;