Add logging to file and compile time macro to change log level
Added compile time flag (LOG_MAX_LEVEL) that can be used to change the
applications logging level at compile time.

Also implemented the logging to file policy so we can log to a file
if we choose.

Updated main to show an example of how to set the file stream.